임시 토큰 설정
